+++ R/manip_transform.R
@@ -506,3 +506,16 @@ to_no_isolates.network <- function(.data) {
 to_no_isolates.data.frame <- function(.data) {
   as_edgelist(to_no_isolates(as_tidygraph(.data)))
 }
+
+#' @describeIn transform Galois derivations
+#' @export
+to_galois <- function(.data) {
+  x <- as_matrix(.data)
+  thisRequires("multiplex")
+  out <- multiplex::galois(x, labeling = "reduced")
+  out <- multiplex::partial.order(out, type = "galois")
+  class(out) <- c("matrix", class(out))
+  rownames(out)[!startsWith(rownames(out), "{")] <- ""
+  colnames(out)[!startsWith(colnames(out), "{")] <- ""
+  out
+}